Add initial benchmarks

+++ src/Prometheus/Metric/Counter.hs
@@ -9,10 +9,11 @@ import Prometheus.Metric
 import Prometheus.Metric.TVar
 import Prometheus.MonadMetric
 
+import Data.Int (Int64)
 import qualified Control.Concurrent.STM as STM
 
 
-newtype Counter = MkCounter (STM.TVar Integer)
+newtype Counter = MkCounter (STM.TVar Int64)
 
 counter :: Info -> MetricGen Counter
 counter info = do
@@ -22,7 +23,7 @@ counter info = do
         ,   collect = collectTVar info CounterType valueTVar
         }
 
-withCounter :: MonadMetric m => Metric Counter -> (Integer -> Integer) -> m ()
+withCounter :: MonadMetric m => Metric Counter -> (Int64 -> Int64) -> m ()
 withCounter (Metric {handle = MkCounter valueTVar}) f =
     doIO $ STM.atomically $ STM.modifyTVar' valueTVar f
